// packages/api-client/src/generated/apis/AccountAuthControllerApi.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime';
import {
    type LoginResponse,
    LoginResponseFromJSON,
    LoginResponseToJSON,
} from '../models/LoginResponse';
import {
    type PhoneSmsAuthRequest,
    PhoneSmsAuthRequestFromJSON,
    PhoneSmsAuthRequestToJSON,
} from '../models/PhoneSmsAuthRequest';

export interface PhoneSmsAuthOperationRequest {
    phoneSmsAuthRequest: PhoneSmsAuthRequest;
}

/**
 * 
 */
export class AccountAuthControllerApi extends runtime.BaseAPI {

    /**
     * Creates request options for phoneSmsAuth without sending the request
     */
    async phoneSmsAuthRequestOpts(requestParameters: PhoneSmsAuthOperationRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['phoneSmsAuthRequest'] == null) {
            throw new runtime.RequiredError(
                'phoneSmsAuthRequest',
                'Required parameter "phoneSmsAuthRequest" was null or undefined when calling phoneSmsAuth().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';


        let urlPath = `/api/v1/accounts/phone-sms-auth`;

        return {
            path: urlPath,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: PhoneSmsAuthRequestToJSON(requestParameters['phoneSmsAuthRequest']),
        };
    }

    /**
     */
    async phoneSmsAuthRaw(requestParameters: PhoneSmsAuthOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<LoginResponse>> {
        const requestOptions = await this.phoneSmsAuthRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => LoginResponseFromJSON(jsonValue));
    }

    /**
     */
    async phoneSmsAuth(requestParameters: PhoneSmsAuthOperationRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<LoginResponse> {
        const response = await this.phoneSmsAuthRaw(requestParameters, initOverrides);
        return await response.value();
    }

}