// packages/api-client/src/generated/apis/AccountProfileControllerApi.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime';
import {
    type AccountProfileResponse,
    AccountProfileResponseFromJSON,
    AccountProfileResponseToJSON,
} from '../models/AccountProfileResponse';
import {
    type UpdateDisplayNameRequest,
    UpdateDisplayNameRequestFromJSON,
    UpdateDisplayNameRequestToJSON,
} from '../models/UpdateDisplayNameRequest';

export interface PatchMeRequest {
    updateDisplayNameRequest: UpdateDisplayNameRequest;
}

/**
 * 
 */
export class AccountProfileControllerApi extends runtime.BaseAPI {

    /**
     * Creates request options for getMe without sending the request
     */
    async getMeRequestOpts(): Promise<runtime.RequestOpts> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};


        let urlPath = `/api/v1/accounts/me`;

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     */
    async getMeRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<AccountProfileResponse>> {
        const requestOptions = await this.getMeRequestOpts();
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => AccountProfileResponseFromJSON(jsonValue));
    }

    /**
     */
    async getMe(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<AccountProfileResponse> {
        const response = await this.getMeRaw(initOverrides);
        return await response.value();
    }

    /**
     * Creates request options for patchMe without sending the request
     */
    async patchMeRequestOpts(requestParameters: PatchMeRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['updateDisplayNameRequest'] == null) {
            throw new runtime.RequiredError(
                'updateDisplayNameRequest',
                'Required parameter "updateDisplayNameRequest" was null or undefined when calling patchMe().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';


        let urlPath = `/api/v1/accounts/me`;

        return {
            path: urlPath,
            method: 'PATCH',
            headers: headerParameters,
            query: queryParameters,
            body: UpdateDisplayNameRequestToJSON(requestParameters['updateDisplayNameRequest']),
        };
    }

    /**
     */
    async patchMeRaw(requestParameters: PatchMeRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<AccountProfileResponse>> {
        const requestOptions = await this.patchMeRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => AccountProfileResponseFromJSON(jsonValue));
    }

    /**
     */
    async patchMe(requestParameters: PatchMeRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<AccountProfileResponse> {
        const response = await this.patchMeRaw(requestParameters, initOverrides);
        return await response.value();
    }

}